+++ src/Makefile.am
@@ -92,6 +92,7 @@ BITCOIN_CORE_H = \
   compressor.h \
   consensus/consensus.h \
   consensus/params.h \
+  consensus/validation.h \
   core_io.h \
   eccryptoverify.h \
   ecwrapper.h \
