+++ src/rpcblockchain.cpp
@@ -4,7 +4,9 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "checkpoints.h"
+#include "consensus/validation.h"
 #include "main.h"
+#include "primitives/transaction.h"
 #include "rpcserver.h"
 #include "sync.h"
 #include "util.h"
