+++ src/txmempool.cpp
@@ -7,6 +7,7 @@
 
 #include "clientversion.h"
 #include "consensus/consensus.h"
+#include "consensus/validation.h"
 #include "main.h"
 #include "policy/fees.h"
 #include "streams.h"
