+++ src/wallet/wallet.cpp
@@ -9,6 +9,7 @@
 #include "checkpoints.h"
 #include "coincontrol.h"
 #include "consensus/consensus.h"
+#include "consensus/validation.h"
 #include "main.h"
 #include "net.h"
 #include "script/script.h"
