+++ src/wallet/walletdb.cpp
@@ -6,6 +6,7 @@
 #include "wallet/walletdb.h"
 
 #include "base58.h"
+#include "consensus/validation.h"
 #include "main.h"
 #include "protocol.h"
 #include "serialize.h"
